+++ web/static/react/assets/ColdStartInterviewPage-CICcvHyR.js
@@ -1,0 +1,1 @@
+import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as t}from"./LegalSkillsProfilePage-oBFa-w1X.js";import"./button-BT6xASFN.js";import"./index-B280S8W7.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-nwKQIl4I.js";import"./IusPageShell-DUvk0hIK.js";import"./api-BB02Po9l.js";function n(){return r.jsx(t,{})}export{n as ColdStartInterviewPage};