+++ web/static/react/assets/PracticeProfilePage-CH_LH9p1.js
@@ -1,0 +1,1 @@
+import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as o}from"./LegalSkillsProfilePage-oBFa-w1X.js";import"./button-BT6xASFN.js";import"./index-B280S8W7.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-nwKQIl4I.js";import"./IusPageShell-DUvk0hIK.js";import"./api-BB02Po9l.js";function l(){return r.jsx(o,{})}export{l as PracticeProfilePage};