+++ web/static/react/assets/PracticeProfilePage-CTvnzBSE.js
@@ -1,1 +1,0 @@
-import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as o}from"./LegalSkillsProfilePage-CCzZEqux.js";import"./button-DJhBvu1Y.js";import"./index-BvWUaJfh.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-DbaF0qtG.js";import"./IusPageShell-Cvcw1w6o.js";import"./api-DuIksswa.js";function l(){return r.jsx(o,{})}export{l as PracticeProfilePage};